Bunch o' stuff

+++ src/main/java/com/foxobyte/pokenary/repo/WildPokemonRepository.java
@@ -1,7 +1,7 @@
 package com.foxobyte.pokenary.repo;
 
-import com.foxobyte.pokenary.dao.WildPokemon;
+import com.foxobyte.pokenary.dao.pokemon.WildPokemon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WildPokemonRepository extends JpaRepository<WildPokemon, Integer> {
+public interface WildPokemonRepository extends JpaRepository<WildPokemon, Long> {
 }
